Naviplus/backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
wheel==0.45.1
